misc: copy folders to operations/ (#6080)
This change copies folders from `production/` to `operations/` that we
intend to keep (the static mode mixin and operator related manifests).

I tried to originally make symlinks and move the folders, but this doesn't 
work in the GitHub UI, where it fails to navigate to a file within a symlinked 
directory. Since there are references to such files, symlinks won't work for 
us here.

Once references to the productions folder are completely eliminated, the 
folder will be removed.

Related to #6077.