flow/internal/controller: wire up references between DeclareNodes
This commit updates the Loader to detect and add edges between
DeclareNodes for DeclareNodes which depend on each other. This works by
recursively looking for component definitions within a DeclareNode's AST
that reference a component defined by a DeclareNode.

Note that only top-level declare blocks are backed by DeclareNodes:

  declare "a" {
    declare "a_1" { }
  }

In the above example, a DeclareNode is created for `a` but not for
`a_1`. Because of this, the dependencies of a declare block are bubbled
up to the root DeclareNode. This can lead to misleading error messages:

  declare "a" {
    declare "a_1" {
      b "example" { }
    }

    a_1 "example" { }
  }

  declare "b" {
    a "example" { }
  }

In the above example, there is a cyclic dependency between a, a_1, and
b. However, the user receives an error reporting a cycle _just_ between
a and b, as a_1's dependency on b bubbles up to a from the perspective
of the DAG.

This is not an ideal error message, but is arguably acceptable for an
initial implementation.