flow/internal/controller: introduce CustomComponentNode
CustomComponentNode represents a custom component in the DAG. A
CustomComponentNode creates a "module" (in the old terminology) which it
loads with the definition of a DeclareNode and a set of arguments
provided to it.

CustomComponentNode is currently not used, but is added in an
independent commit due to the volume of code introduced.