flow/internal/controller: support creating custom components
This change updates the Loader to create instances of
CustomComponentNode.

The Loader now offers a ComponentRegistry of current DeclareComponents
found in its graph. To support this change, the constructed graph is now
stored before evaluation. The Flow controller uses this custom
ComponentRegistry to propagate available custom components down to inner
controllers.