Add CustomComponent interface
CustomComponent is defined in the controller and aimed to be managed by CustomComponentNodes.
This replaces the component.Module interface which will be removed with the old modules.
This change aligns with the new modules terminology and keeps the custom components within the controller package.
A custom component can load a config via LoadBody and run via Run.
LoadOptions are added to allow users to pass additional options to the loader with the new config.
For now, the LoadOptions contain a CustomComponentRegistry which provides custom component templates that the custom component can use to instantiate other custom components.