Add ComponentNodeManager
The ComponentNodeManager is responsible for creating new components nodes and obtaining the necessary information to run them.
The data is stored in a builtinComponentRegistry and a customComponentRegistry.
The created custom components must be wired to the declare and import nodes that they depend on to ensure that their config is loaded
before they evaluate and to trigger a re-evaluation when an import node they depend on received new content.