Add ImportConfigNode
Add a new config node that retrieves a module via an import source.
It will parse the module to collect declare and import blocks.
For every imported import block it will create ImportConfigNode children.
The children are evaluated and ran by the parent.
By navigating through the import tree, it's possible to access all imported declares.