Update flow controller to use the ImportConfigNode.
When an ImportConfigNode is evaluated, the custom component registry is updated.
Custom components have edges to import node that they depend on.
The componentNodeManager can search in the customComponentRegistry for imported declares.
ImportConfigNodes are runnable nodes. They are ran by the controller.
When a custom component is instantiated with an imported declare, it will have the customComponentRegistry associated with the scope of the import.