introduce framework for OpenTelemetry Collector config conversions
This commit introduces a general framework for converting OpenTelemetry
Collector configuration files into Flow mode configuration files.

The framework relies on implementations of the componentConverter
interface which convert a single OpenTelemetry Collector component into
one or more Flow components.

For this initial commit, conersion for receivers, processors, and
exporters are supported. Conversion for connectors and extensions will
need to be made in future commits.