flowtest: introduce script-based testing framework for Flow
This commit introduces a new package, flow/flowtest, which can be used
for running script-based test using Flow components.

flowtest accepts a txtar archive, where the comment of the txtar archive
denotes the script to run. Files in the txtar archive are unpacked to a
temporary directory as part of the test; that temporary directory is
also used as the working directory of the Flow controller.

See <rsc.io/script> for more information about the scripting language
and available default commands.

In addition to the default commands, the following Flow-specific
commands are made available to scripts:

- `controller.load [file]`: Load a file into the Flow controller.
- `controller.start`: Start the Flow controller.
- `assert.health [component] [health]`: Assert the health of a specific
  component.

Note that `controller.start` should almost always be run as a background
command to allow the controller to run for the duration of the test.

Users of flowtest can provide custom commands on top of these by passing
`WithExtraCommands` to `TestScript`.

The following is an example script using these commands:

  # This file performs a basic test of loading a component and asserting its
  # health.

  controller.load main.river
  controller.start &

  assert.health local.file.example healthy

  -- main.river --
  local.file "example" {
    filename = "./hello.txt"
  }

  -- hello.txt --
  Hello, world!

See the testdata/ directory for other test examples.