Add framework for converting OpenTelemetry Collector configs into Flow mode configs (#6379)
* introduce framework for OpenTelemetry Collector config conversions

This commit introduces a general framework for converting OpenTelemetry
Collector configuration files into Flow mode configuration files.

The framework relies on implementations of the componentConverter
interface which convert a single OpenTelemetry Collector component into
one or more Flow components.

For this initial commit, conersion for receivers, processors, and
exporters are supported. Conversion for connectors and extensions will
need to be made in future commits.

* otelcolconvert: add stringify utilities

This commit adds stringify utilities for converting OpenTelemetry
instance IDs and River builder.Blocks to strings. This is intended to be
used by implementations of componentConverter for adding info-level
diagnostics.

* otelcolconvert: add converters for otlpreceiver, otlpexporter

This commit adds converters for both the otlpreceiver and oltpexporter.
OpenTelemetry Collector requires at least one receiver and one exporter,
so both are needed to be able to produce a basic working example of
config conversion.

* converter: ignore empty lines in diagnostics

Some editors don't remove trailing newlines from files, causing blank
lines to appear as an expected diagnostic.