Make *auth.Handler into an interface
This is so that we can extend the interface with a `RiverTokenize`
method to wire in during the conversion. Alternatively, we could
introduce the RiverTokenize method on the existing implementation, but
first, it would be hard to reason about why it's required by a totally
different package, and second, its implementation could only be called
in a useful way from the otelconvert package.

Signed-off-by: Paschalis Tsilias <paschalist0@gmail.com>