dockertarget: fix flaky test (#6625)
The callback for one of the assert.EventuallyWithT calls incorrectly
passed the test-wide *testing.T instance. This causes that
assert.EventuallyWithT call to always return on the first tick,
regardless of whether the assertion passed.

This means that on slower machines, the first assertion would fail, the
assert.EventuallyWithT call would return prematurely, and the rest of
the test would fail in turn.