otelcol.receiver.opencensus: sync with upstream defaults
This commit re-syncs `otelcol.receiver.opencensus` with the upstream
defaults, namely changing the default listen port from 4317 to 55678.

This is considered a breaking change; a section has been written in the
upgrade guide notifying users to explicitly set the endpoint to the
previous port to retain old behavior when upgrading.

Closes #6489.