docs: exempt converters from backwards compatibility guarantees
Converter code is not cheap to maintain, and in some cases can bring in
many dependencies that are otherwise unused, making project maintainance
more difficult.

This change explicitly marks converter code as exempt from backwards
compatibility guarntees, allowing its removal between minor versions if
a converter is no longer needed and maintainers don't want to wait for
the next major release to remove it.

The RFC for backwards compatibility guarantees is also updated to make
it clear that some functionality may opt out of backwards compatibility
guarantees, which includes the `convert` and `tools` commands.