packaging: properly unregister Grafana Agent when uninstalling on Windows (#166)
grafana/agent#6550 fixed an issue where the service settings were not
being removed properly, but inadvertedly broke the process of removing
the application from the Add/Remove programs list.

Grafana AGent is registered as a program to the 32-bit registry, and so it must
be removed from the 32-bit registry and not the 64-bit registry.