[grafana/agent] Format with `prettier`
Consistent formatting improves readability and makes it easier for tools to transform the source.

The general human understanding of Markdown has some ambiguities so it's possible this PR will break some documentation presentation because the formatter follows the CommonMark specification.

In this case, this PR specifically enables a follow up PR where links are automatically updated to follow the current guidance in Writers' Toolkit.
Consider adopting `prettier` in your local workflow and enforcing it in CI to help eliminate those ambiguities in the future.

Repository: grafana/agent
Created-By: migrator